LaCosaNostra ejecutada con FIFO sin Threads.

+++ demo/src/main/java/com/example/IPiatto.java
@@ -1,7 +1,7 @@
 package com.example;
 
 public interface IPiatto {
-    float tempoDiCottura = 0;
+    long tempoDiCottura = 0;
 
-    public float getTempoDiCottura();
+    public long getTempoDiCottura();
 }
